requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
pycoingecko==3.1.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
pydantic<2.0.0
python-multipart==0.0.6
aiohttp==3.9.1
